Update README.md
This makes more clear the nature of the disclosures required.  It also clarifies that the disclosures must be made prior to the promise being made, so that a party might withdraw a nomination in the event that the disclosures reveal a material problem.  This version deletes the consent condition because whereas the promise is made to both parties, consent to an appointment is not required by both.